backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
python-dotenv==1.0.0
itsdangerous==2.1.2
bcrypt==4.0.1
email-validator==2.1.0.post1
marshmallow==3.20.1
